Unificar tipo MemeItem para versiones multiidioma
